pyproject.toml
[project]
name = "ece-genait"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.118.0",
    "langchain>=0.3.27",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.78",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langchain-tavily>=0.1.5",
    "langgraph>=0.6.8",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langgraph-cli[inmem]>=0.4.2",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.1",
    "accelerate>=0.30.0",
    "sentence-transformers>=5.2.2",
    "transformers>=4.40.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.37.0",
    "watchdog>=6.0.0",
    "streamlit>=1.54.0",
    "requests>=2.32.5",
    "httpx>=0.28.1",
]